refactor: unify article generation
