+++ src/worker.ts
@@ -1,14 +1,8 @@
 import server from './_worker.js';
 import cron from './cron-worker';
-import { generateArticle } from './modules/articleGenerator';
-import { generateHeroImage } from './modules/heroImageGenerator';
-import { publishArticleToGitHub } from './modules/githubPublisher';
-import articleTemplate from './prompt/article-content.txt?raw';
-import heroTemplate from './prompt/hero-image.txt?raw';
-import { getRecentTitlesFromGitHub } from './utils/recentTitlesGitHub';
+import { generateAndPublish } from './modules/generateAndPublish';
 import { initLogger, logRequest, logEvent, logError } from './utils/logger';
 import { getSessionInfo, appendSessionCookie } from './utils/session';
-import { slugify } from './utils/slugify';
 import { escapeHtml } from './utils/escapeHtml';
 
 async function handleContact(request: Request, env: Env) {
@@ -48,20 +42,6 @@ async function handleContact(request: Request, env: Env) {
   }
 }
 
-async function generateAndPublish(env: Env) {
-  const recent = await getRecentTitlesFromGitHub(env.GITHUB_REPO, env.GITHUB_TOKEN);
-  const recentList = recent.map((t, i) => `${i + 1}. ${t}`).join('\n');
-  const articlePrompt = articleTemplate.replace('{recent_titles}', recentList);
-  const article = await generateArticle({
-    apiKey: env.OPENAI_API_KEY,
-    prompt: articlePrompt,
-    maxTokens: 7200,
-  });
-  const heroPrompt = heroTemplate.replace('{title}', article.title);
-  const heroImage = await generateHeroImage({ apiKey: env.OPENAI_API_KEY, prompt: heroPrompt });
-  await publishArticleToGitHub({ env, article, heroImage });
-  return { article, slug: slugify(article.title) };
-}
 
 async function handleGenerateArticleJson(request: Request, env: Env) {
   logEvent({ type: 'generate-article-endpoint-start' });
